refactor cart elements
